# perf(proxy/compute): Improve HTML Document parsing