### Checklist

* [x] I have read the [Contributor Guide](../../CONTRIBUTING.md)
* [x] I have read and agree to the [Code of Conduct](../../CODE_OF_CONDUCT.md)
* [x] I have added a description of my changes and why I'd like them included in the section below

### Description of Changes

This PR improves the HTML Document parsing by doing the following:

- Stop parsing the document when we have extracted all the data we need (mainly: SDK tag, data layer tag if exists, `<title>`, canonical and keywords meta tags)
- Considers the `<head>`-only tags as non-defined/empty when they're not present

The "trade-off" for this improved parsing function is that the Edgee SDK tag *must* be present early in the document, since that's the main "exit point" of the function.

## Benchmarks

Benchmark results using https://github.com/edgee-cloud/edgee-benches repository:
```shell
$ cargo criterion base.*/printemps
    Finished `bench` profile [optimized + debuginfo] target(s) in 0.08s

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

process-html/base/printemps-bagues-femme.html (1046.1 KB)                                                                             
                        time:   [3.1260 ms 3.1583 ms 3.1949 ms]
                        thrpt:  [312.27 MiB/s 315.89 MiB/s 319.15 MiB/s]
process-html/base-improv/printemps-bagues-femme.html (1046.1 KB)                                                                            
                        time:   [108.02 µs 109.18 µs 110.24 µs]
                        thrpt:  [8.8378 GiB/s 8.9237 GiB/s 9.0193 GiB/s]
process-html/base/printemps-avant-premiere-black-friday.html (1.7 MB)                                                                            
                        time:   [5.4944 ms 5.5695 ms 5.6628 ms]
                        thrpt:  [291.99 MiB/s 296.88 MiB/s 300.94 MiB/s]
process-html/base-improv/printemps-avant-premiere-black-friday.html (1.7 MB)                                                                            
                        time:   [143.82 µs 144.86 µs 146.15 µs]
                        thrpt:  [11.048 GiB/s 11.147 GiB/s 11.227 GiB/s]
process-html/base/printemps-recherche.html (2.3 MB)                                                                            
                        time:   [7.0625 ms 7.2004 ms 7.3514 ms]
                        thrpt:  [301.92 MiB/s 308.25 MiB/s 314.27 MiB/s]
process-html/base-improv/printemps-recherche.html (2.3 MB)                                                                            
                        time:   [166.22 µs 167.13 µs 168.14 µs]
                        thrpt:  [12.891 GiB/s 12.969 GiB/s 13.040 GiB/s]
```

Note that these results are from running the code on my dev laptop on a mostly-recent CPU